Move props types and values to the right spots
