+++ front/app/component-library/components/CheckboxWithLabel/index.tsx
@@ -2,48 +2,35 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import { Color } from '../../utils/styleUtils';
 import testEnv from '../../utils/testUtils/testEnv';
 import Box, { BoxMarginProps, BoxPaddingProps } from '../Box';
-import Checkbox from '../Checkbox';
+import Checkbox, { CheckboxProps } from '../Checkbox';
 import IconTooltip from '../IconTooltip';
 
 const CheckboxContainer = styled.div<{ hasLabel: boolean }>`
   margin-right: ${({ hasLabel }) => (hasLabel ? '10px' : '0px')};
 `;
 
-type DefaultProps = {
-  size?: string;
-  disabled?: boolean;
-  indeterminate?: boolean;
-};
-
-type Props = DefaultProps &
+type Props = {
+  label?: string | JSX.Element | null;
+  labelTooltipText?: string | JSX.Element | null;
+  dataTestId?: string;
+} & CheckboxProps &
   BoxPaddingProps &
-  BoxMarginProps & {
-    checked: boolean;
-    onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
-    label?: string | JSX.Element | null;
-    labelTooltipText?: string | JSX.Element | null;
-    id?: string;
-    name?: string;
-    stopLabelPropagation?: boolean;
-    checkedColor?: Color;
-    dataTestId?: string;
-  };
+  BoxMarginProps;
 
 const CheckboxWithLabel = ({
   id,
+  size,
+  disabled,
+  indeterminate,
   label,
-  labelTooltipText,
   stopLabelPropagation,
-  size = '24px',
   checked,
-  disabled = false,
-  indeterminate = false,
   onChange,
   name,
   checkedColor,
+  labelTooltipText,
   dataTestId,
   ...boxProps
 }: Props) => {
@@ -52,8 +39,6 @@ const CheckboxWithLabel = ({
   const handleLabelClick = (event: React.MouseEvent) => {
     stopLabelPropagation && event.stopPropagation();
   };
-
-  const checkedOrIndeterminate = checked || indeterminate;
 
   return label ? (
     <Box
@@ -73,7 +58,6 @@ const CheckboxWithLabel = ({
           checked={checked}
           disabled={disabled}
           checkedColor={checkedColor}
-          checkedOrIndeterminate={checkedOrIndeterminate}
           indeterminate={indeterminate}
           size={size}
           name={name}
@@ -91,7 +75,6 @@ const CheckboxWithLabel = ({
       checked={checked}
       disabled={disabled}
       checkedColor={checkedColor}
-      checkedOrIndeterminate={checkedOrIndeterminate}
       indeterminate={indeterminate}
       size={size}
       name={name}
